Move entities to assets
