+++ frontend/assets/entities/application.entity.ts
