+++ frontend/assets/entities/event.entity.ts
