+++ frontend/assets/entities/settings.entity.ts
