+++ frontend/components/Main.tsx
@@ -2,10 +2,10 @@ import React, { ReactElement, useEffect, useMemo } from "react";
 import { createStyles, Grid, makeStyles, Typography } from "@material-ui/core";
 import axios, { AxiosResponse } from "axios";
 
-import { Configuration } from "./entities/configuration.entity";
+import { Configuration } from "../assets/entities/configuration.entity";
 import { defaultConfiguration } from "./defaultConfiguration";
 import { parsedQuery, useSettings } from "./Utils";
-import { Setting } from "./entities/settings.entity";
+import { Setting } from "../assets/entities/settings.entity";
 import ConfigurationComponent from "./Configuration/Configuration";
 import Logs from "./Logs/Logs";
 import WebRTC from "./WebRTC/WebRTC";
